+++ src/main/resources/initDB.sql
@@ -27,6 +27,6 @@ CREATE TABLE dishes (
   id            INTEGER PRIMARY KEY DEFAULT nextval('global_seq'),
   restaurant_id INTEGER   NOT NULL,
   description   TEXT      NOT NULL,
-  price      FLOAT       NOT NULL,
+  price      INTEGER       NOT NULL,
   FOREIGN KEY (restaurant_id) REFERENCES restaurants (id) ON DELETE CASCADE
 );
